% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_report_format.R
\name{fetch_report_format}
\alias{fetch_report_format}
\title{Fetch The Specified Report Formatting Information}
\usage{
fetch_report_format(cfg, rptname = "default", element = "Default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name initialized with \code{system_report_init}}

\item{element}{report element to fetch: for Word reports it can be 
"default" (default),       "Normal",        "Code",  "TOC",          
"Heading_1",  "Heading_2", "Heading_3", "Table", "Table_Labels",
"Table_Caption", "Figure",  and  "Figure_Caption"}
}
\value{
list of current parameter gauesses
}
\description{
Returns a list the default font format for the report element
}
