% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_rpttype.R
\name{fetch_rpttype}
\alias{fetch_rpttype}
\title{Determines Type of Report Template}
\usage{
fetch_rpttype(
  template = file.path(system.file(package = "onbrand"), "templates", "report.pptx")
)
}
\arguments{
\item{template}{Name of PowerPoint or Word file}
}
\value{
List with the following elements
\itemize{
\item{rpttype} Either Word, PowerPoint or Unknown
\item{isgood} Boolean variable indicating success or failure
\item{msgs} Vector of messages
}
}
\description{
Based on the file extension for a template
}
\examples{
fetch_rpttype(template= file.path(system.file(package="onbrand"), "templates", "report.pptx"))

}
