% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_apply_md.R
\name{ft_apply_md}
\alias{ft_apply_md}
\title{Render Markdown in flextable Object}
\usage{
ft_apply_md(ft, obnd = NULL, part = "body")
}
\arguments{
\item{ft}{Flextable object.}

\item{obnd}{Optional onbrand object used to format markdown. The default \code{NULL} value will use default formatting.}

\item{part}{Part of the table can be one of \code{"all"}, \code{"body"} (default), \code{"header"}, or \code{"footer"}}
}
\value{
flextable with markdown applied
}
\description{
Takes a flextable object and renders any markdown in the
specified part.
}
