% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_to_oo.R
\name{md_to_oo}
\alias{md_to_oo}
\title{Parse Markdown into Officer as_paragraph Result}
\usage{
md_to_oo(strs, default_format = NULL)
}
\arguments{
\item{strs}{vector of strings containing Markdown can contain the following elements:}

\item{default_format}{list containing the default format for elements not defined with markdown default values (format the same as \code{\link{md_to_officer}}, default is \code{NULL})}
}
\value{
list with the following elements
\itemize{
 \item{isgood}:    Boolean value indicating the result of the function call
 \item{msgs}:      sequence of strings contianing a description of any problems
 \item{as_par_cmd}:as_paragraph generated code from md_to_officer
 \item{oo}:        as_paragraph officer object resulting from running the as_par_cmd code
}
}
\description{
Used to take small markdown chunks and return the as_paragraph
results. This function will take the markdown specified in str, calls
md_to_officer, evals the as_paragraph field from the first paragraph
returned, evals that result and returns the object from the as_paragraph
command.
}
\examples{
res = md_to_oo("Be **bold**")
}
