% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onbrand.R
\docType{package}
\name{onbrand}
\alias{onbrand-package}
\alias{onbrand}
\title{onbrand: officer Abstraction Layer for Organizational Templates}
\description{
The onbrand package creates an abstraction layer that is easily
configurable with a yaml file to allow for creation of reproducible
reporting work flows across Word and PowerPoint templates.
}
\seealso{
\url{https://github.com/john-harrold/onbrand}
}
\author{
\strong{Maintainer}: John Harrold \email{john.m.harrold@gmail.com} (\href{https://orcid.org/0000-0003-2052-4373}{ORCID})

Authors:
\itemize{
  \item Bryan Smith \email{r.bryan.smith@gmail.com}
}

}
