% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_alignment.R
\name{plot_genome_alignment}
\alias{plot_genome_alignment}
\title{Plot Genome Alignment}
\usage{
plot_genome_alignment(
  q_name,
  r_name,
  fq,
  fr,
  ref_Ns,
  query_Ns,
  plot_df,
  sz = 0.009,
  line = FALSE,
  ns = TRUE,
  gaps = NULL,
  min_mapQ = 0,
  gapwidth = 1e+05,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{q_name}{A character string representing the query name.}

\item{r_name}{A character string representing the reference name.}

\item{fq}{A vector of query sequences.}

\item{fr}{A vector of reference sequences.}

\item{ref_Ns}{A data frame containing reference sequence information.}

\item{query_Ns}{A data frame containing query sequence information.}

\item{plot_df}{A data frame containing the plot data.}

\item{sz}{A numeric value representing the size of the points or lines in the plot (default: 0.009).}

\item{line}{A logical value indicating whether to display alignments as lines (default: FALSE).}

\item{ns}{A logical value indicating whether to display reference and query sequence lines (default: TRUE).}

\item{gaps}{A data frame containing gap information, or NULL if not provided (default: NULL).}

\item{min_mapQ}{A numeric value representing the minimum mapping quality for filtering (default: 0).}

\item{gapwidth}{A numeric value representing the minimum gap width (default: 1e5).}

\item{xlim}{A numeric vector representing the x-axis limits, or NULL if not provided (default: NULL).}

\item{ylim}{A numeric vector representing the y-axis limits, or NULL if not provided (default: NULL).}
}
\value{
A ggplot object representing the genome alignment plot.
}
\description{
This function creates a visual representation of genome alignments using ggplot2.
It allows for customization of the plot, including filtering by mapping quality, displaying gaps, and setting axis limits.
}
\examples{
# Load sample data and call the plot_genome_alignment() function
# (Note: Replace with your own data when using this function)
plot_genome_alignment(q_name = "Query", r_name = "Target", fq = fq_data, fr = fr_data,
ref_Ns = ref_Ns_data, query_Ns = query_Ns_data, plot_df = plot_df_data)
}
